refactor(tools): 重命名 verify_image 为 read_image，文件名对齐工具名
