+++ tools/__init__.py
@@ -48,7 +48,7 @@ def get_all_tools() -> list[BaseTool]:
     from tools.network.tool_weather import WeatherTool
     from tools.network.tool_holiday import HolidayCalendarTool
     from tools.network.tool_image_understand import ImageUnderstandTool
-    from tools.network.tool_verify_image import ReadImageTool
+    from tools.network.tool_read_image import ReadImageTool
     from tools.network.tavily import TavilySearchTool, TavilyExtractTool
 
     # Files
